+++ pkg/gatewayserver/io/io.go
@@ -62,6 +62,8 @@ type Server interface {
 	ClaimDownlink(ctx context.Context, ids ttnpb.GatewayIdentifiers) error
 	// UnclaimDownlink releases the claim of the downlink path for the given gateway.
 	UnclaimDownlink(ctx context.Context, ids ttnpb.GatewayIdentifiers) error
+	// FromRequestContext decouples the lifetime of the provided context from the values found in the context.
+	FromRequestContext(ctx context.Context) context.Context
 }
 
 // Connection is a connection to a gateway managed by a frontend.
@@ -438,6 +440,7 @@ func (c *Connection) ScheduleDown(path *ttnpb.DownlinkPath, msg *ttnpb.DownlinkM
 			eirp = *sb.MaxEIRP
 		}
 		settings := ttnpb.TxSettings{
+			DataRate:      dr.Rate,
 			DataRateIndex: rx.dataRateIndex,
 			Frequency:     rx.frequency,
 			Downlink: &ttnpb.TxSettings_Downlink{
@@ -448,7 +451,6 @@ func (c *Connection) ScheduleDown(path *ttnpb.DownlinkPath, msg *ttnpb.DownlinkM
 		if int(ids.AntennaIndex) < len(c.gateway.Antennas) {
 			settings.Downlink.TxPower -= c.gateway.Antennas[ids.AntennaIndex].Gain
 		}
-		settings.DataRate = dr.Rate
 		if lora := dr.Rate.GetLoRa(); lora != nil {
 			settings.CodingRate = phy.LoRaCodingRate
 			settings.Downlink.InvertPolarization = true