+++ pkg/webui/console/components/events/previews/application-uplink.js
@@ -26,10 +26,11 @@ import JSONPayload from './shared/json-payload'
 const ApplicationUplinkPreview = React.memo(({ event }) => {
   const { data, identifiers } = event
   const deviceIds = identifiers[0].device_ids
-  let snr
+  let snr, rssi
 
   if ('rx_metadata' in data) {
     snr = data.rx_metadata[0].snr
+    rssi = data.rx_metadata[0].rssi
   }
 
   const bandwidth = getByPath(data, 'settings.data_rate.lora.bandwidth')
@@ -49,6 +50,7 @@ const ApplicationUplinkPreview = React.memo(({ event }) => {
       )}
       <DescriptionList.Item title={messages.fPort} data={data.f_port} />
       <DescriptionList.Item title={messages.snr} data={snr} />
+      <DescriptionList.Item title={messages.rssi} data={rssi} />
       <DescriptionList.Item title={messages.bandwidth} data={bandwidth} />
     </DescriptionList>
   )